extend EditMode and move some files from model into new util folder
